Reverse the order of hwaddr update and master update
If team devices and port ownership are setup before teamd is started
the change of master for a given port will be detected before the
hardware address is detected. This is problematic since storing of the
original hardware address is guarded by master not being set. This patch
reverses this order so that hardware address change is detected before
master change is checked as stored. This is important for ie. the LACP
runner which uses the original hardware address of the ports when
sending LACPDU frames.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>