ttdp-lw: Fix port state for team ports added before teamd is started
If a team port is already owned by the team interface before starting teamd we
need to sync the port state when starting, since the port wont automatically
get a port-change event which would trigger the sync otherwise.

This patch forces a sync of port state when the port is initilized.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>